+++ src/schema.js
@@ -3,9 +3,9 @@ const { gql } = require('apollo-server');
 const typeDefs = gql`
   type Race {
     id: Int!
-    distance: Int!
-    startTime: Int
-    endTime: Int
+    distance: Float!
+    startTime: Float
+    endTime: Float
     users: [User]
   }
 
@@ -30,7 +30,7 @@ const typeDefs = gql`
 
   type Score {
     id: Int!
-    time: Int!
+    time: Float!
     user: User!
   }
 
@@ -44,7 +44,7 @@ const typeDefs = gql`
   }
 
   type Mutation {
-    createRace(distance: Int!, startTime: Int, endTime: Int): Race!
+    createRace(distance: Float!, startTime: Float, endTime: Float): Race!
     createUser(username: String!, position: Int, RaceId: Int!): User!
     createLocation(
       startLat: Float!
@@ -54,10 +54,10 @@ const typeDefs = gql`
       distance: Float!
       UserId: Int!
     ): Location!
-    createScore(time: Int!, UserId: Int!): Score!
+    createScore(time: Float!, UserId: Int!): Score!
     updateUser(id: Int!, position: Int!): User!
-    updateRaceStartTime(id: Int!, startTime: Int!): Race!
-    updateRaceEndTime(id: Int!, endTime: Int!): Race!
+    updateRaceStartTime(id: Int!, startTime: Float!): Race!
+    updateRaceEndTime(id: Int!, endTime: Float!): Race!
     updateLocation(
       id: Int!
       endLat: Float!
