+++ test/resolvers/user.test.js
@@ -10,7 +10,7 @@ const models = require('../../models');
 const db = require('../../models/index');
 
 const CREATE_RACE = gql`
-  mutation createRace($distance: Int!) {
+  mutation createRace($distance: Float!) {
     createRace(distance: $distance) {
       id
       distance
@@ -68,7 +68,7 @@ const CREATE_LOCATION = gql`
   }
 `;
 const CREATE_SCORE = gql`
-  mutation createScore($time: Int!, $UserId: Int!) {
+  mutation createScore($time: Float!, $UserId: Int!) {
     createScore(time: $time, UserId: $UserId) {
       id
       time
